Add a Camera that can point in different directions
The 'origin' is the position of the camera in the scene.

The 'direction' is where the camera is pointing.

The 'orientation' specifies which direction is up.

The middle of the film will be placed at the tail of the direction
vector. For example, if th camera's origin is (1, 1, 1) and its
direction is (0, 0, 1) then the middle of the film is (1, 1, 2).